core: refactored normalizers
